+++ hyp3_autorift/process.py
@@ -261,6 +261,7 @@ def apply_wallis_nodata_fill_filter(array: np.ndarray, nodata: int) -> Tuple[np.
 
     array, _ = prepare_array_for_filtering(array, nodata)
     filtered, zero_mask = _wallis_filter_fill(array, filter_width=5, std_cutoff=0.25)
+    filtered[zero_mask] = 0
 
     return filtered, zero_mask
 
@@ -273,7 +274,7 @@ def _apply_filter_function(image_path: str, filter_function: Callable) -> Tuple[
 
     image_new_path = create_filtered_filepath(image_path)
     _ = io.write_geospatial(image_new_path, image_filtered, image_transform, image_projection,
-                            nodata=0, dtype=gdal.GDT_Float32)
+                            nodata=None, dtype=gdal.GDT_Float32)
 
     zero_path = None
     if zero_mask is not None:
@@ -387,6 +388,9 @@ def process(reference: str, secondary: str, parameter_file: str = DEFAULT_PARAME
 
         filter_platform = min([platform, get_platform(secondary)])
         if filter_platform in ('L4', 'L5', 'L7'):
+            # Log path here before we transform it
+            log.info(f'Reference scene path: {reference_path}')
+            log.info(f'Secondary scene path: {secondary_path}')
             reference_path, reference_zero_path, secondary_path, secondary_zero_path = \
                 apply_landsat_filtering(reference_path, secondary_path)