+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.10.5]
+
+### Fixed
+* The zero mask and nodata value for wallis-filtered Landsat-7 images are now set appropriately
+* Early (SLC-On) Landsat-7 images are no longer incorrectly filtered a second time with the high-pass filter
+* The search range and %-valid pixels are now correctly calculated for Landsat-7 images
+
 ## [0.10.4]
 
 ### Fixed